+++ src/components/ui_list/ModalEmployee.tsx
@@ -1,6 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { deleteEmployee, employeeClose, setActiveEmployee } from '../../redux/actions/employeeAction';
 import { useNavigate } from 'react-router-dom';
+import Swal from 'sweetalert2';
 export const ModalEmployee = () => {
     const {activeEmployee} = useSelector((state: any) => state.employees);
     const dispatch = useDispatch();
@@ -14,7 +15,20 @@ export const ModalEmployee = () => {
 
     const handleDelete = (e: any) => {
         e.preventDefault();
-        dispatch(deleteEmployee(uid));
+        Swal.fire({
+            title: '¿Desea borrar el empleado?',
+            text: "¡No podrá deshacer esta acción!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Borrar',
+            cancelButtonText: 'Cancelar'
+        }).then((result) => {
+            if (result.value) {
+                dispatch(deleteEmployee(uid));
+            }
+        })
     }
 
     const handleEdit = (e: any) =>{
@@ -28,8 +42,10 @@ export const ModalEmployee = () => {
         <div className="h-5/6 w-10/12 md:w-10/12 lg:w-9/12 -translate-y-0 mr-2 sm:mr-4 md:mr-8">
             <div className="bg-green-600 shadow-lg hidden xs:inline-block w-full h-full rounded-3xl absolute bottom-0 transform -rotate-3"></div>
         
-            <div className="bg-white transition shadow-xl w-full h-full rounded-3xl absolute flex flex-col items-start px-6 xs:px-10 sm:px-16 md:px-14 py-6 md:py-10">
-                <div className=' relative z-50 w-full flex flex-row justify-end items-center mb-0.5 md:mb-2' >
+            <div className="bg-white transition  shadow-xl w-full h-full rounded-3xl absolute flex flex-col items-start px-6 xs:px-10 sm:px-16 md:px-14 py-6 md:py-10">
+                <div className=' relative z-50 w-full flex flex-row justify-between items-center mb-0.5 md:mb-2' >
+                    <h1 className='text-gray-800 text-2xl font-bold'>Información Empleado </h1>
+                    
                     <button title='cerrar' className=' w-8 xs:w-8 md:w-12 h-8 xs:h-8 md:h-12 rounded-full border-2 border-solid border-green-800'
                     onClick={(e) =>setClose(e)} >
                         <i className="fa-solid fa-x text-rojoPokemon"></i>
@@ -37,7 +53,7 @@ export const ModalEmployee = () => {
                 </div>
             </div>
 
-            <div className='relative p-20'>
+            <div className='relative p-20 mt-6'>
                 <div className = "flex flex-col">
                     <h1 className='text-gray-800 text-xl font-bold'>Nombre Completo: </h1>
                     <div className='flex flex-row'>
@@ -46,8 +62,8 @@ export const ModalEmployee = () => {
                     </div>
                 </div>
 
-                <div className='flex flex-row mt-6'>
-                    <div className='flex- flex-col mr-20' >
+                <div className='flex flex-row mt-6 w-8/12 justify-between'>
+                    <div className='flex- flex-col' >
                         <h1 className='text-gray-800 text-xl font-bold'>Correo Electrónico: </h1>
                         <h1 className='text-gray-600 text-lg font-bold'>{email}</h1>
                     </div>
@@ -58,19 +74,19 @@ export const ModalEmployee = () => {
                     </div>
                 </div>
 
-                <div className='flex flex-row mt-6'>
-                    <div className='flex flex-col mr-20' >
+                <div className='flex flex-row mt-6 w-8/12 justify-between'>
+                    <div className='flex flex-col' >
                         <h1 className='text-gray-800 text-xl font-bold'>Fecha de Nacimiento: </h1>
                         <h1 className='text-gray-600 text-lg font-bold'>{fechaNacimiento}</h1>
                     </div>
 
-                    <div className='flex flex-col mr-20' >
+                    <div className='flex flex-col' >
                         <h1 className='text-gray-800 text-xl font-bold'>Número de cédula: </h1>
                         <h1 className='text-gray-600 text-lg font-bold'>{cedula}</h1>
                     </div>
                 </div>
                 
-                <div className='w-full flex flex-row justify-between mt-10 px-16' >
+                <div className='w-full flex flex-row justify-between mt-16 px-16' >
                       <button onClick={(e) => handleEdit(e)} 
                       title="Editar" className="bg-blue-500 rounded-full w-16 h-16 mx-1 text-white hover:shadow-lg text-lg">
                         <i className="fa-solid fa-user-pen text-2xl"></i>
